JDF compliance
