+++ examples/errors/pom.xml
@@ -1,43 +1,59 @@
+<?xml version="1.0"?>
 <!--
-  Copyright (C) Red Hat, Inc.
-  http://redhat.com
+  JBoss, Home of Professional Open Source
+   Copyright 2013, Red Hat, Inc. and/or its affiliates, and individual
+   contributors by the @authors tag. See the copyright.txt in the
+   distribution for a full listing of individual contributors.
 
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+   http://www.apache.org/licenses/LICENSE-2.0
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
   -->
+
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-         
+
     <modelVersion>4.0.0</modelVersion>
 
-    <!--
-      The parent POM does not contain any information required to build this example, it's only here
-      for organisational purposes so feel free to remove/replace this.
-    -->
-    <parent>
-        <groupId>org.jboss.fuse.examples</groupId>
-        <artifactId>project</artifactId>
-        <version>99-master-SNAPSHOT</version>
-    </parent>
-
-    <groupId>org.jboss.fuse.examples</groupId>
+    <groupId>org.jboss.quickstarts.fuse</groupId>
     <artifactId>errors</artifactId>
+    <version>99-master-SNAPSHOT</version>
     <packaging>jar</packaging>
     <name>JBoss Fuse :: Examples :: Error Handling</name>
     <description>Handle errors in the Camel m</description>
 
     <properties>
+        <!-- Explicitly declaring the source encoding eliminates the following
+                   message: -->
+        <!-- [WARNING] Using platform encoding (UTF-8 actually) to copy filtered
+            resources, i.e. build is platform dependent! -->
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+
+
+        <version.camel-blueprint>2.10.0.redhat-60024</version.camel-blueprint>
+        <version.slf4j>1.6.1</version.slf4j>
+        <version.junit>4.10</version.junit>
+
+        <version.compiler.plugin>2.3.1</version.compiler.plugin>
+
+        <!-- maven-compiler-plugin -->
+        <maven.compiler.target>1.6</maven.compiler.target>
+        <maven.compiler.source>1.6</maven.compiler.source>
     </properties>
+
+    <licenses>
+        <license>
+            <name>Apache License, Version 2.0</name>
+            <distribution>repo</distribution>
+            <url>http://www.apache.org/licenses/LICENSE-2.0.html</url>
+        </license>
+    </licenses>
 
     <dependencies>
         <!--
@@ -48,7 +64,7 @@
         <dependency>
             <groupId>org.apache.camel</groupId>
             <artifactId>camel-blueprint</artifactId>
-            <version>${camel-version}</version>
+            <version>${version.camel-blueprint}</version>
             <scope>provided</scope>
         </dependency>
         <!--
@@ -57,7 +73,7 @@
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
-            <version>${slf4j-api-version}</version>
+            <version>${version.slf4j}</version>
             <scope>provided</scope>
         </dependency>
 
@@ -68,7 +84,7 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>${junit-version}</version>
+            <version>${version.junit}</version>
             <scope>test</scope>
         </dependency>
         <!--
@@ -77,7 +93,7 @@
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-log4j12</artifactId>
-            <version>${slf4j-api-version}</version>
+            <version>${version.slf4j}</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
@@ -108,7 +124,7 @@
     </repositories>
 
     <pluginRepositories>
-         <pluginRepository>
+        <pluginRepository>
             <id>fusesource.m2</id>
             <name>FuseSource Community Release Repository</name>
             <url>http://repo.fusesource.com/nexus/content/repositories/releases</url>
@@ -119,7 +135,7 @@
                 <enabled>true</enabled>
             </releases>
         </pluginRepository>
-         <pluginRepository>
+        <pluginRepository>
             <id>fusesource.ea</id>
             <name>FuseSource Community Early Access Release Repository</name>
             <url>http://repo.fusesource.com/nexus/content/groups/ea</url>
@@ -132,4 +148,18 @@
         </pluginRepository>
     </pluginRepositories>
 
+    <build>
+
+        <plugins>
+            <!-- Compiler plugin enforces Java 1.6 compatibility and activates annotation processors -->
+            <plugin>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>${version.compiler.plugin}</version>
+                <configuration>
+                    <source>${maven.compiler.source}</source>
+                    <target>${maven.compiler.target}</target>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
 </project>
