+++ examples/errors/src/main/resources/data/order4.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  JBoss, Home of Professional Open Source
+   Copyright 2013, Red Hat, Inc. and/or its affiliates, and individual
+   contributors by the @authors tag. See the copyright.txt in the
+   distribution for a full listing of individual contributors.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+   http://www.apache.org/licenses/LICENSE-2.0
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+  -->
+
+<order xmlns="http://fusesource.com/examples/order/v7">
+
+    <customer id="D0004">
+        <name>Dartmoor Zoological Park</name>
+        <city>Devon</city>
+        <country>UK</country>
+    </customer>
+
+    <date>2012-03-04</date>
+
+    <orderlines>
+        <orderline>
+            <article id="D0040">
+                <description>Dinosaur</description>
+            </article>
+            <quantity>4</quantity>
+        </orderline>
+        <orderline>
+            <article id="D0400">
+                <description>Dragonfly</description>
+            </article>
+            <quantity>40</quantity>
+        </orderline>
+    </orderlines>
+</order>
