--- examples/errors/src/test/data/order1.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<order xmlns="http://fusesource.com/examples/order/v7">
-
-    <customer id="A0001">
-        <name>Antwerp Zoo</name>
-        <city>Antwerp</city>
-        <country>BE</country>
-    </customer>
-
-    <date>2012-03-01</date>
-
-    <orderlines>
-        <orderline>
-            <article id="A0001">
-                <description>Aardvark</description>
-            </article>
-            <quantity>1</quantity>
-        </orderline>
-        <orderline>
-            <article id="A0011">
-                <description>Alpaca</description>
-            </article>
-            <quantity>10</quantity>
-        </orderline>
-    </orderlines>
-
-</order>
