--- examples/errors/src/test/data/order2.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<order xmlns="http://fusesource.com/examples/order/v7">
-
-    <customer id="B0002">
-        <name>Bristol Zoo Gardens</name>
-        <city>Bristol</city>
-        <country>UK</country>
-    </customer>
-
-    <date>2012-03-02</date>
-
-    <orderlines>
-        <orderline>
-            <article id="B0002">
-                <description>Badger</description>
-            </article>
-            <quantity>2</quantity>
-        </orderline>
-        <orderline>
-            <article id="B0202">
-                <description>Bee</description>
-            </article>
-            <quantity>200</quantity>
-        </orderline>
-    </orderlines>
-</order>
